Use correct NPM icon for Node-related files
This is the same 16×16-sized icon that the NPM registry uses, so it only
makes sense to use it in Atom too. Since this *technically* isn't Node's
icon, the internal class names have been updated to say "npm" instead of
"node" (this is trivial, but felt obliged to note it all the same).

The previous NPM icon is still retained in the icon file for anybody who
wishes to revert back to the original in their own stylesheets.

Resolves #211: Update NPM icon.