Refactor logic for defining Directory instances
This shall be the final commit related to the filesystem API. All future
work will be handled in a separate codebase. Most people would agree the
sheer size of this API has long since transcended the scope of a package
meant for file-icons. Indeed, this code is closer to something one would
find in Atom's core, tabs notwithstanding.

Notable changes made in this commit:

* The `Resource` class is now representative of an actual filesystem
  object. It was originally intended as an abstract base-class only;
  however, recent developments have highlighted a need for something
  more consistent and defined.

  It's probable the ArchiveEntry class will go the way of TreeEntry
  and be made a standalone class (without being a Resource subclass).

* The `.repo` property is now global to all resources, and pertains
  to the enclosing repository. On a similar note, `isRepository` is
  now used for determining if a Directory matches a Git repository.

* Filesystem stats are now the sole indicator of what Resource type
  something is. This brings the underlying logic closer to symmetry
  with the filesystem, although the handling of symbolic links will
  be revised in future.