Decouple utility functions from package codebase
Adding a hard dependency on a library of helper functions is something I
wouldn't do until it was seriously needed. We've now reached that point:
these functions are needed by both the package and the filesystem API.

Moving them all to the filesystem module would feel hacky and weird, and
it'd also pollute its entry point with numerous unrelated exports.