Add --auto switch to executable icon-compiler
The package's config was originally planned to be overridden by users in
the form of a user directory resource: ~/.atom/file-icons.cson. A script
was written to enable reading and writing to arbitrary config locations.

The idea of user-controlled configs was later ditched, as it would offer
users nothing they can't already do using their stylesheets. This script
remained in case recompilation would ever be needed outside of Atom.