Add icon for Pod documents and change Perl's icon
The camel belongs to O'Reilly Publishing, while the onion is the logo of
the Perl Foundation. It's therefore closer to an "official" logo for the
language... and incidentally, it happens to make a much nicer icon, too.

Users who prefer the camel icon can add this to their stylesheets:

    .perl-icon::before{
        content: "\f142";
        font-family: Mfizz;
        font-size: 14px;
        top: 1px;
    }

Trivia for non-Perl folk:
    https://en.wikipedia.org/wiki/Pearl_onion#Cultural_references