Use test-icons when matching hashbangs
Although one can use their stylesheet to fix incorrect icons, doing this
quickly becomes laborious when too many permutations are involved:

    [data-name*="-spec."][data-name$=".py"]::after,
    [data-name*="-test."][data-name$=".py"]::after,
    [data-name*="-specs."][data-name$=".py"]::after,
    [data-name*="-tests."][data-name$=".py"]::after{
        font: 17px file-icons;
        content: "\ea66";
        top: 3px;
    }

This commit is a stopgap for an oft-encountered scenario where tests are
given inconsistent icons because they contain hashbangs. Of course, this
only solves part of a much larger issue with package's inability to link
multiple icons to a language, as opposed to only one. That's better left
to a future rewrite, however.

Fixes: file-icons/atom#812